requirements.txt
certifi==2017.4.17
chardet==3.0.4
coreapi==2.3.1
coreschema==0.0.4
Django==1.11.3
djangorestframework==3.6.3
djangorestframework-jwt==1.11.0
idna==2.5
itypes==1.1.0
Jinja2==2.9.6
MarkupSafe==1.0
PyJWT==1.5.2
pytz==2017.2
requests==2.18.1
uritemplate==3.0.0
urllib3==1.21.1
